CrossOrigin Added

+++ src/main/java/com/maveric/userservice/service/UserServiceImpl.java
@@ -17,6 +17,9 @@ import org.springframework.stereotype.Service;
 import java.util.ArrayList;
 import java.util.List;
 
+import static com.maveric.userservice.constants.Constants.USER_DELETED_SUCCESS;
+import static com.maveric.userservice.constants.Constants.USER_NOT_FOUND_MESSAGE;
+
 @Service
 public class UserServiceImpl implements UserService {
     @Autowired
@@ -30,51 +33,68 @@ public class UserServiceImpl implements UserService {
     public List<UserDto> getUsers(Integer page, Integer pageSize) {
         Pageable paging = PageRequest.of(page, pageSize);
         Page<User> pageResult = repository.findAll(paging);
-        if(pageResult.hasContent()) {
+        if (pageResult.hasContent()) {
             List<User> users = pageResult.getContent();
             return mapper.mapToDto(users);
         } else {
             return new ArrayList<>();
         }
     }
+
+    @Override
     public UserDto createUser(UserDto userDto) {
         String pass = passwordEncoder.encode(userDto.getPassword());
         userDto.setPassword(pass);
-        System.out.println(pass);
-        User user = mapper.map(userDto);
-        User userResult = repository.save(user);
-        return  mapper.map(userResult);
+        User userResult = repository.findByEmail(userDto.getEmail());
+        if (userResult == null) {
+            User user = mapper.map(userDto);
+            User userResult2 = repository.save(user);
+            return mapper.map(userResult2);
+        } else {
+            throw new UserNotFoundException("User Already Exist! for this emailId");
+        }
     }
+
     @Override
     public UserDto getUserDetails(String userId) {
-        User userResult=repository.findById(userId).orElseThrow(() -> new UserNotFoundException("User not found with id " + userId));
+        User userResult = repository.findById(userId).orElseThrow(() -> new UserNotFoundException("User not found with id " + userId));
         return mapper.map(userResult);
     }
+
     @Override
     public String deleteUser(String userId) {
+        if (!repository.findById(userId).isPresent()) {
+            throw new UserNotFoundException(USER_NOT_FOUND_MESSAGE + userId);
+        }
         repository.deleteById(userId);
-        return "User deleted successfully.";
+        return USER_DELETED_SUCCESS;
     }
 
     @Override
     public UserDto getUserDetailsByEmail(String emailId) {
-        User userResult=repository.findByEmail(emailId);
-        return mapper.map(userResult);
+        User userResult = repository.findByEmail(emailId);
+        if (userResult != null)
+            return mapper.map(userResult);
+        else
+            throw new UserNotFoundException("User not found! fir this emailId");
     }
-
     @Override
     public UserDto updateUser(String userId, UserDto userDto) {
-        User userResult=repository.findById(userId).orElseThrow(() -> new UserNotFoundException("User not found"));
-        userResult.set_id(userResult.get_id());
-        userResult.setFirstName(userDto.getFirstName());
-        userResult.setLastName(userDto.getLastName());
-        userResult.setMiddleName(userDto.getMiddleName());
-        userResult.setPhoneNumber(userDto.getPhoneNumber());
-        userResult.setEmail(userDto.getEmail());
-        userResult.setAddress(userDto.getAddress());
-        userResult.setDateOfBirth(userDto.getDateOfBirth());
-        userResult.setGender(userDto.getGender());
-        User accountUpdated = repository.save(userResult);
-        return mapper.map(accountUpdated);
+        if (userId.equals(userDto.get_id())) {
+            User userResult = repository.findById(userId).orElseThrow(() -> new UserNotFoundException("User not found"));
+            userResult.set_id(userResult.get_id());
+            userResult.setFirstName(userDto.getFirstName());
+            userResult.setLastName(userDto.getLastName());
+            userResult.setMiddleName(userDto.getMiddleName());
+            userResult.setPhoneNumber(userDto.getPhoneNumber());
+            userResult.setEmail(userDto.getEmail());
+            userResult.setAddress(userDto.getAddress());
+            userResult.setDateOfBirth(userDto.getDateOfBirth());
+            userResult.setGender(userDto.getGender());
+            User accountUpdated = repository.save(userResult);
+            return mapper.map(accountUpdated);
+        } else {
+            throw new UserNotFoundException("User Id not found! Cannot Update account.");
+        }
     }
 }
